moved purify cast efficiency BrewCDR => PurifyingBrew
