+++ src/parser/monk/brewmaster/modules/features/checklist/Module.js
@@ -39,7 +39,7 @@ class Checklist extends Analyzer {
           isb: this.isb.uptimeSuggestionThreshold,
           purifyHeavy: this.pb.purifyHeavySuggestion,
           purifyDelay: this.pb.purifyDelaySuggestion,
-          purifyCasts: this.brewcdr.purifySuggestionThreshold,
+          purifyCasts: this.pb.purifyCastSuggestion,
           totalCDR: this.brewcdr.suggestionThreshold,
           isbClipping: this.isb.clipSuggestionThreshold,
           bof: this.bof.suggestionThreshold,
